+++ lib/scout_apm/config.rb
@@ -27,6 +27,7 @@ module ScoutApm
         'stackprof_interval' => 20000, # microseconds, 1000 = 1 millisecond, so 20k == 20 milliseconds
         'uri_reporting' => 'full_path',
         'report_format' => 'json',
+        'disabled_instruments' => []
     }.freeze
 
     def initialize(config_path = nil)