+++ CHANGELOG.markdown
@@ -1,3 +1,13 @@
+# 1.4.5
+
+* Instrument Elasticsearch
+* Instrument InfluxDB
+
+# 1.4.4
+
+* Instrument Mongoid
+* Instrument Redis
+
 # 1.4.3
 
 * Add a to_s call to have HTTPClient work with URI objects.  Thanks to Nicolai for providing the change!